Fragmented Author- and CheepService classes into several files.
