+++ src/Chirp.Infrastructure/AuthorService/AuthorViewModel.cs
@@ -1,0 +1,3 @@
+﻿namespace Chirp.Infrastructure.AuthorService;
+
+public record AuthorViewModel(int Id, string Name, string Email);
