+++ src/Chirp.Infrastructure/CheepService/CheepViewModel.cs
@@ -1,0 +1,3 @@
+﻿namespace Chirp.Infrastructure.CheepService;
+
+public record CheepViewModel(string Author, string Message, string Timestamp);
