+++ src/Chirp.Infrastructure/CheepService/ICheepService.cs
@@ -1,0 +1,11 @@
+﻿using Chirp.Core.Models;
+
+namespace Chirp.Infrastructure.CheepService;
+
+public interface ICheepService
+{
+    public Task<PagedResult<CheepViewModel>> GetCheeps(int page, int pageSize);
+    public Task<PagedResult<CheepViewModel>> GetCheepsFromAuthor(string author, int page, int pageSize);
+    public Task<CheepViewModel> GetCheepById(int id);
+    public Task<int> PostCheep(CheepViewModel cheep);
+}
