+++ test/Chirp.InfrastructureTest/CheepServiceUnitTest.cs
@@ -1,5 +1,6 @@
 using Chirp.Core.DataTransferObject;
 using Chirp.Core.Models;
+using Chirp.Infrastructure.CheepService;
 
 namespace Chirp.InfrastructureTest;
